changed scan cb index type to size_t

+++ code/public/memlib/module.hpp
@@ -10,7 +10,7 @@
 
 namespace memlib
 {
-	using multiscan_callback = memlib::address(*)(memlib::address addr, uint32_t index);
+	using multiscan_callback = memlib::address(*)(memlib::address addr, size_t index);
 
 	class module
 	{
